Title: 나눗셈의 나머지, Time: 136ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 251210/나눗셈의 나머지/remainder-of-division.java
@@ -14,8 +14,8 @@ public class Main {
 
         for (int i = 0; i < 100; i++) {
 
-            a = a/b;
-            tmp = a%b;
+            a = (int) a / b;
+            tmp = (int) a % b;
             arr[tmp] += 1;
 
             if(a == 0){
